Build with -Wall -Wconversion; add narrow<T> for checked casts
Enable -Wall -Wconversion for GCC/Clang builds to flag implicit
narrowing and sign-changing conversions the compiler can prove
potentially lossy.  Introduce a narrow<To>(From) template in narrow.h
that does static_cast<To>(x) plus an assertion that the round-trip is
lossless and the sign (or lack thereof) is preserved.  This is NOT a
substitute for validating untrusted input -- it's an assertion that
the surrounding code has reasoned about the value range.  A failed
assertion here is a bug in the caller, not a recoverable error.

Fixes a real bug caught by the new warnings: der_integer() was typed
`long`, and tsa.cpp was calling `der_integer(long(nonce))` to pass a
63-bit random nonce.  On Windows long is 32 bits, which silently
truncated the nonce to half its entropy.  Changed der_integer to take
int64_t and removed the lossy cast.

Other warning fixes (all known-safe narrowing):
- base64.cpp: narrow<uint32_t>(v) where v is an int already filtered >= 0.
- x509.cpp: narrow<size_t>(nbytes) where nbytes is 1..4 after validation.
- pe.cpp: int pad -> size_t pad (the int narrowing was gratuitous since
  the source was uint64_t % 8); read_le16 now computes in uint32_t and
  narrows once at the end to avoid int-promotion warnings.
- posix.cpp: narrow<size_t>(ret) at sites where a prior `if (ret < 0)
  throw` has already excluded negatives.

Also removed the #pragma comment(lib, ...) directives in win32.cpp.
They were MSVC-only and redundant with the existing CMake
target_link_libraries entries (which we need anyway for MinGW).  Keeps
CMake as the single source of truth for link dependencies.

+++ src/base64.cpp
@@ -1,4 +1,5 @@
 #include "base64.h"
+#include "narrow.h"
 
 static const char table[] =
     "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789+/";
@@ -38,7 +39,7 @@ std::vector<uint8_t> base64_decode(const std::string &s)
     for (char c : s) {
         int v = decode_char(c);
         if (v < 0) continue;
-        accum = (accum << 6) | v;
+        accum = (accum << 6) | narrow<uint32_t>(v);
         bits += 6;
         if (bits >= 8) {
             bits -= 8;
